Fix #1518: Add additional check in IsMinimalEndpoint to prevent false positives
When Minimal Apis are not used, and controllers are, ODataQueryOptions<TEntity> always calls PopulateMetadata, which will always add metadata of type ODataMiniMetadata to the endpoint.
Currently, when IsMinimalEndpoint is called, it checks whether ODataMiniMetadata exists which is always the case.
This commits makes it additionally check for an HttpMethodAttribute in the endpoint metadata.